requirements.txt
aiogram-dialog
aiogram
python-dotenv
pydantic-settings
pydantic
sqlalchemy
psycopg2
redis
xlsxwriter